Add a shell wrapper for c.py
This is an interesting script. It enables using c.py with
automatic paging (the built in click stuff was crap). Futhermore
it is now possible to provide *all* cat options, as we can simply
pipe c.py's output through cat. That converts c.py to a full drop
in replacement for cat (because cat is used internally). :)